Make ColumnarBatch extendable
